catalogue: replace sqlite with ripgrep for speed
Searching for a book with a sql query like this took approximately 23
seconds, scanning about 38 million rows.

```sql
select *
from openlibrary
where type = '/type/work'
  and data ->> $.title like '%Provenzalische%'
```

Replacing it with a call to ripgrep, followed by a verification that
the match was actually in the title (or id) lie this takes approximately
2 seconds.

```
rg --fixed-strings "Provenzalische" .data/openlibrary/ol_dump_works_2025-02-11.txt
```

So yeah, why bother with a database?