import.goodreads: read the `reviews.json` to figure out read dates
This isn't a perfect system; the `reviews.json` file you get from an
account export doesn't contain all the reviews for a single book; just
the "current" review.  In other words, if you have read a book 4 times,
there is one review entry, and it has a create date of when the book was
added, and an updated of the most recent read date.